refactor: add parser utils for reusable functions
