+++ packages/nimble-components/src/rich-text/editor/testing/rich-text-editor.pageobject.ts
@@ -3,6 +3,7 @@ import type { RichTextEditor } from '..';
 import { waitForUpdatesAsync } from '../../../testing/async-helpers';
 import type { ToggleButton } from '../../../toggle-button';
 import type { ToolbarButton } from './types';
+import { getTagsFromDocumentFragment, getLeafContentsFromDocumentFragment } from '../../models/testing/markdown-parser-utils';
 
 /**
  * Page object for the `nimble-rich-text-editor` component.
@@ -123,17 +124,11 @@ export class RichTextEditorPageObject {
     }
 
     public getEditorTagNames(): string[] {
-        return Array.from(this.getTiptapEditor()!.querySelectorAll('*')).map(
-            el => el.tagName
-        );
+        return getTagsFromDocumentFragment(this.getTiptapEditor() as HTMLElement);
     }
 
     public getEditorLeafContents(): string[] {
-        return Array.from(this.getTiptapEditor()!.querySelectorAll('*'))
-            .filter((el, _) => {
-                return el.children.length === 0;
-            })
-            .map(el => el.textContent || '');
+        return getLeafContentsFromDocumentFragment(this.getTiptapEditor() as HTMLElement);
     }
 
     public getFormattingButtonTextContent(
